Made skalo command commensurable with the file input logic that runs build if you provide a list of fasta files
