+++ src/skalo_utils.rs
@@ -1,6 +1,5 @@
 use hashbrown::{HashMap, HashSet};
-use std::fs::File;
-use std::io::Write;
+use std::str::FromStr;
 
 pub fn rev_compl(seq: &str) -> String {
     let reverse_code: HashMap<char, char> =
@@ -65,7 +64,7 @@ pub fn encode_kmer(kmer: &str) -> u128 {
     result
 }
 
-pub fn decode_kmer(encoded: u128, k: usize) -> String {
+pub fn decode_kmer_skalo(encoded: u128, k: usize) -> String {
     let bits_to_nucleotide: [char; 4] = ['A', 'C', 'G', 'T'];
     let mut kmer = String::with_capacity(k);
 
@@ -129,58 +128,7 @@ impl<'a> VariantInfo<'a> {
     }
 }
 
-pub fn identify_good_kmers(
-    len_kmer: usize,
-    all_kmers: &HashMap<u128, HashMap<u128, u32>>,
-    index_map: &HashMap<u32, String>,
-) -> (HashSet<u128>, HashSet<u128>) {
-    println!(" # identify bubble extremities");
-
-    let len_kmer_graph = len_kmer - 1;
-
-    let mut start_kmers: HashSet<u128> = HashSet::new();
-    let mut end_kmers: HashSet<u128> = HashSet::new();
-
-    for (kmer, next_kmers_map) in all_kmers.iter() {
-        if next_kmers_map.len() > 1 {
-            let all_next_kmer: Vec<&u128> = next_kmers_map.keys().collect();
-
-            'i_loop: for (i, &next_kmer) in all_next_kmer.iter().enumerate() {
-                for &next_kmer2 in all_next_kmer.iter().skip(i + 1) {
-                    let samples_1 = &index_map[&next_kmers_map[next_kmer]];
-                    let samples_2 = &index_map[&next_kmers_map[next_kmer2]];
-
-                    if compare_samples(samples_1, samples_2) {
-                        start_kmers.insert(kmer.clone());
-
-                        let dna = decode_kmer(kmer.clone(), len_kmer_graph);
-                        let rc = rev_compl(&dna);
-
-                        //uncomment to print network
-                        //println!("{}	{}	red", &dna, &dna);
-
-                        end_kmers.insert(encode_kmer(&rc));
-
-                        //uncomment to print network
-                        //println!("{}	{}	red", &rc, &rc);
-                        break 'i_loop;
-                    }
-                }
-            }
-        }
-    }
-
-    // exit program if no extremity found (e.g. cases of weeded skf files)
-    if start_kmers.is_empty() {
-        eprintln!("\n      Error: there is no entry node in this graph, hence no variant.\n");
-        std::process::exit(1);
-    }
-
-    println!("     . {} entry nodes", start_kmers.len());
-    (start_kmers, end_kmers)
-}
-
-fn compare_samples(str_1: &str, str_2: &str) -> bool {
+pub fn compare_samples(str_1: &str, str_2: &str) -> bool {
     let set1: HashSet<&str> = str_1.split('|').collect();
     let set2: HashSet<&str> = str_2.split('|').collect();
     let diff_1 = set1.difference(&set2).count();
@@ -189,241 +137,7 @@ fn compare_samples(str_1: &str, str_2: &str) -> bool {
     diff_1 != 0 && diff_2 != 0
 }
 
-pub fn filter_output_sequences(
-    variant_groups: HashMap<String, Vec<VariantInfo>>,
-    len_kmer: usize,
-    all_samples: Vec<String>,
-    n_homopolymer: Option<u32>,
-    max_missing: f32,
-    output_name: &str,
-    input_name: &str,
-) {
-    println!(" # filter sequences");
-
-    let len_kmer_graph = len_kmer - 1;
-
-    // prepare output files
-    let mut out_1 =
-        File::create(format!("{}_seq_groups.fas", output_name)).expect("Failed to create file");
-    let mut out_2 =
-        File::create(format!("{}_variants.tsv", output_name)).expect("Failed to create file");
-
-    // write header of TSV variant file
-    out_2
-        .write_all(format!("#skalo version {}\n", env!("CARGO_PKG_VERSION")).as_bytes())
-        .expect("Failed to write to file");
-    out_2
-        .write_all(
-            format!(
-                "#parameters: m={}{}\n",
-                max_missing,
-                match n_homopolymer {
-                    Some(n) => format!(" n={}", n),
-                    None => String::new(),
-                }
-            )
-            .as_bytes(),
-        )
-        .expect("Failed to write to file");
-    out_2
-        .write_all(format!("#input file: {} (k={})\n", input_name, len_kmer).as_bytes())
-        .expect("Failed to write to file");
-
-    let index_name: Vec<String> = all_samples
-        .iter()
-        .enumerate()
-        .map(|(index, value)| format!("{}:{}", index, value))
-        .collect();
-    let formatted_string = index_name.join(", ");
-    out_2
-        .write_all(format!("#samples: {}\n", formatted_string).as_bytes())
-        .expect("Failed to write to file");
-    out_2.write_all(b"#pos_ali\tnb_states\ttype\tunclear_insert\tfirst_kmer\tvariants\tlast_kmer\tratio_missing\tsamples\n")
-        .expect("Failed to write to file");
-
-    // Prepare variables for binary alignment
-    let mut binary_seq: HashMap<String, Vec<String>> = HashMap::new();
-    let mut d_samples: HashMap<String, String> = HashMap::new();
-    for (i, sample) in all_samples.iter().enumerate() {
-        binary_seq.insert(sample.clone(), Vec::new());
-        d_samples.insert(i.to_string(), sample.clone()); // map sample ID to sample full name
-    }
-
-    let mut nb_missing = 0;
-    let mut nb_homopolymer = 0;
-
-    // variant groups 1 by 1
-    let mut position = 0;
-    for vec_variants in variant_groups.values() {
-        if vec_variants[0].is_snp {
-
-            // do something with SNPs
-        } else {
-            // check the ratio of missing samples
-            let ratio_missing = calculate_ratio_missing(all_samples.len(), vec_variants.clone());
-
-            if ratio_missing > max_missing {
-                nb_missing += 1;
-            } else {
-                // CHECK IF UNCLEAR INSERT
-                // extract 1st kmer using 1st sequence
-                let first_kmer = vec_variants[0].sequence[..len_kmer_graph].to_string();
-
-                // collect 'middle-bases' and last k-mer (forward and reverse-complement)
-                let (l_middle_bases, last_kmer) =
-                    collect_middle_bases(vec_variants.clone(), len_kmer_graph, false);
-                let (rc_l_middle_bases, rc_last_kmer) =
-                    collect_middle_bases(vec_variants.clone(), len_kmer_graph, true);
-
-                // test if multiple positions exist for the last k-mer
-                let multiple_positions =
-                    test_multiple_positions(l_middle_bases.clone(), last_kmer.clone());
-                let multiple_positions_rc =
-                    test_multiple_positions(rc_l_middle_bases, rc_last_kmer);
-
-                let mut unclear_insert = "no";
-                if multiple_positions || multiple_positions_rc {
-                    unclear_insert = "yes";
-                }
-
-                // CHECK IK HOMOPOLYMER
-                // check if indel in homopolymer
-                let mut is_homopolymer = false;
-                if let Some(mut n_max) = n_homopolymer {
-                    // adjust n_homopolymer if higher than k-mer length to avoid the program crashing
-                    if n_max > len_kmer_graph.try_into().unwrap() {
-                        n_max = len_kmer_graph as u32;
-                        println!("     . n_homopolymer was reduced to fit the k-mer length");
-                    }
-                    // test presence homopolymer
-                    is_homopolymer =
-                        check_homopolymer(n_max, first_kmer.clone(), l_middle_bases.clone());
-                }
-
-                if is_homopolymer {
-                    nb_homopolymer += 1;
-                } else {
-                    // Output seq_groups
-                    for variant in vec_variants.iter() {
-                        let str_samples = variant
-                            .maj_samples
-                            .iter()
-                            .copied()
-                            .collect::<Vec<_>>()
-                            .join("|");
-                        out_1
-                            .write_all(
-                                format!(">{}_{}\n{}\n", position, str_samples, variant.sequence)
-                                    .as_bytes(),
-                            )
-                            .expect("Failed to write to file");
-                    }
-
-                    // update binary alignment ('-' = missing data; '?' = both states (= 'unknown'))
-                    // only consider the 2 most frequent variants in cases of 3+
-                    let mut sample_done: HashSet<String> = HashSet::new();
-                    let mut state = 0;
-                    for variant in vec_variants.iter().take(2) {
-                        // collect sample names for this sequence and update its vector in binary_seq
-                        for sample_id in &variant.maj_samples {
-                            let full_name = d_samples.get(sample_id.to_owned()).unwrap();
-                            sample_done.insert(full_name.clone());
-                            let sample_vec = binary_seq.get_mut(full_name).unwrap();
-
-                            if sample_vec.len() <= position {
-                                sample_vec.resize(position + 1, state.to_string());
-                            } else {
-                                sample_vec[position] = "?".to_string();
-                            }
-                        }
-                        // update number of character states
-                        state += 1;
-                    }
-
-                    // update binary alignment with missing samples
-                    for sample in &all_samples {
-                        if !sample_done.contains(sample) {
-                            binary_seq
-                                .entry(sample.to_string())
-                                .and_modify(|vec| vec.push("-".to_string()));
-                        }
-                    }
-
-                    // get type of variant
-                    let mut type_variant = "complex";
-                    for xx in &l_middle_bases {
-                        if xx == "." {
-                            type_variant = "_indel_";
-                        }
-                    }
-
-                    // get lists of samples
-                    let mut l_samples = Vec::new();
-                    for variant in vec_variants.iter() {
-                        // sort list of samples by equivalent integers
-                        let mut vect_idx = variant.maj_samples.iter().copied().collect::<Vec<_>>();
-                        vect_idx.sort_by(|a, b| {
-                            let a_int = a.parse::<i32>().unwrap_or(i32::MAX);
-                            let b_int = b.parse::<i32>().unwrap_or(i32::MAX);
-                            a_int.cmp(&b_int)
-                        });
-                        // save it
-                        l_samples.push(vect_idx.join(","));
-                    }
-
-                    // round ratio_missing to the 2nd decimal
-                    let rounded_missing = (ratio_missing * 100.0).round() / 100.0;
-
-                    // save variants information
-                    out_2
-                        .write_all(
-                            format!(
-                                "{}\t{}\t{}\t{}\t{}\t{}\t{}\t{}\t{}\n",
-                                position,
-                                l_middle_bases.len(),
-                                type_variant,
-                                unclear_insert,
-                                first_kmer,
-                                l_middle_bases.join(" / "),
-                                last_kmer,
-                                rounded_missing,
-                                l_samples.join(" / ")
-                            )
-                            .as_bytes(),
-                        )
-                        .expect("Failed to write to file");
-
-                    // Update position
-                    position += 1;
-                }
-            }
-        }
-    }
-
-    // Close output files
-    out_1.flush().expect("Failed to flush file");
-    out_2.flush().expect("Failed to flush file");
-
-    // Output alignment
-    let mut out_3 =
-        File::create(format!("{}_binary_ali.fas", output_name)).expect("Failed to create file");
-    for (sample, l_seq) in &binary_seq {
-        out_3
-            .write_all(format!(">{}\n{}\n", sample, l_seq.join("")).as_bytes())
-            .expect("Failed to write to file");
-    }
-    out_3.flush().expect("Failed to flush file");
-
-    // final message
-    println!("     . {} removed because of missing data", nb_missing);
-    if n_homopolymer.is_some() {
-        println!("     . {} removed because in homopolymers", nb_homopolymer);
-    }
-    println!("     . {} FINAL variant groups", position);
-    println!("done.");
-}
-
-fn calculate_ratio_missing(nb_total: usize, vec_variants: Vec<VariantInfo>) -> f32 {
+pub fn calculate_ratio_missing(nb_total: usize, vec_variants: Vec<VariantInfo>) -> f32 {
     // initialise vector of 0
     let mut ref_samples: Vec<u32> = vec![0; nb_total];
 
@@ -443,7 +157,7 @@ fn calculate_ratio_missing(nb_total: usize, vec_variants: Vec<VariantInfo>) -> f
     ratio_missing
 }
 
-fn collect_middle_bases(
+pub fn collect_middle_bases(
     vec_variants: Vec<VariantInfo>,
     len_k_graph: usize,
     use_rev_complement: bool,
@@ -513,7 +227,7 @@ fn collect_middle_bases(
     (l_middles, last_kmer)
 }
 
-fn test_multiple_positions(l_middles: Vec<String>, last_kmer: String) -> bool {
+pub fn test_multiple_positions(l_middles: Vec<String>, last_kmer: String) -> bool {
     for middle in l_middles {
         let full_string = middle.to_string() + &last_kmer;
 
@@ -533,7 +247,7 @@ fn test_multiple_positions(l_middles: Vec<String>, last_kmer: String) -> bool {
     false // only one occurrence found
 }
 
-fn check_homopolymer(limit_n: u32, first_kmer: String, l_middle_bases: Vec<String>) -> bool {
+pub fn check_homopolymer(limit_n: u32, first_kmer: String, l_middle_bases: Vec<String>) -> bool {
     let mut is_homopolymer = false;
 
     // extract last limit_n nucleotides of first k-mer
@@ -552,4 +266,33 @@ fn check_homopolymer(limit_n: u32, first_kmer: String, l_middle_bases: Vec<Strin
         }
     }
     is_homopolymer
+}
+
+pub fn jaccard_similarity(set1: &HashSet<&str>, set2: &HashSet<&str>) -> f32 {
+    // returns the Jaccard similarity value between 2 sets
+    let intersection_size = set1.intersection(set2).count() as f32;
+    let union_size = (set1.len() as f32 + set2.len() as f32 - intersection_size) as f32;
+    intersection_size / union_size
+}
+
+pub fn count_occurrences(vec: &Vec<u32>) -> HashMap<u32, i32> {
+    let mut count_map = HashMap::new();
+    for num in vec {
+        *count_map.entry(*num).or_insert(0) += 1;
+    }
+    count_map
+}
+
+pub fn convert_combined(input: &str, k: usize) -> Result<String, String> {
+    let parts: Vec<&str> = input.split('@').collect();
+
+    // parse the u128 values from the string parts
+    let u128_1 = u128::from_str(parts[0]).map_err(|e| e.to_string())?;
+    let u128_2 = u128::from_str(parts[1]).map_err(|e| e.to_string())?;
+
+    // compute the reverse complements for both u128 values using the same k
+    let rev_compl_1 = rev_compl_u128(u128_1, k);
+    let rev_compl_2 = rev_compl_u128(u128_2, k);
+
+    Ok(format!("{}@{}", rev_compl_2, rev_compl_1))
 }
